Design: 스타일 공통 요소 추가 및 분리
* default style 에 xxLarge font size 추가
* 공통되는 Border 와 Flex Style common 으로 분리

Related to: #2
